+++ src/test/java/jenkins/internal/provider/SoapXProviderTest.java
@@ -1,9 +1,9 @@
 package jenkins.internal.provider;
 
-import com.sun.xml.internal.messaging.saaj.soap.ver1_2.Message1_2Impl;
 import org.junit.Test;
 
 import javax.xml.soap.SOAPException;
+import javax.xml.soap.SOAPMessage;
 import java.io.IOException;
 
 import static org.mockito.ArgumentMatchers.any;
@@ -13,7 +13,7 @@ public class SoapXProviderTest {
     
     @Test
     public void writeTo() throws SOAPException, IOException {
-        Message1_2Impl soapMessageMock = mock(Message1_2Impl.class);
+        SOAPMessage soapMessageMock = mock(SOAPMessage.class);
         SOAPException exception = mock(SOAPException.class);
         doThrow(exception).when(soapMessageMock).writeTo(any());
