+++ src/test/java/jenkins/task/TestUtil/FakeTask.java
@@ -1,23 +1,20 @@
 package jenkins.task.TestUtil;
 
-import hudson.AbortException;
-import hudson.EnvVars;
 import hudson.Extension;
-import jenkins.internal.data.TestConfiguration;
-import jenkins.task.ExamTask;
+import jenkins.internal.ClientRequest;
+import jenkins.task.Task;
 
-public class FakeTask extends ExamTask {
+import java.io.IOException;
 
-    public FakeTask(String examName, String pythonName, String examReport, String systemConfiguration) {
-        super(examName, pythonName, examReport, systemConfiguration);
+public class FakeTask extends Task {
+    
+    @Override
+    protected void doExecuteTask(ClientRequest clientRequest) throws IOException, InterruptedException {
+    
     }
-
-    protected TestConfiguration addDataToTestConfiguration(TestConfiguration tc, EnvVars env) throws AbortException {
-        return tc;
-    }
-
+    
     @Extension
-    public static class DescriptorFakeTask extends DescriptorExamTask {
-
+    public static class DescriptorFakeTask extends DescriptorTask {
+    
     }
 }
