+++ src/test/java/jenkins/task/TestrunFilterDescriptorTest.java
@@ -7,19 +7,20 @@ import org.junit.Test;
 import org.jvnet.hudson.test.JenkinsRule;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 
 public class TestrunFilterDescriptorTest {
-
+    
     @Rule
     public JenkinsRule jenkinsRule = new JenkinsRule();
-
+    
     private String name;
     private String value;
     private boolean adminCases;
     private boolean activeTestcases;
     private TestrunFilter testObject;
     TestrunFilter.DescriptorImpl testObjectDescriptor;
-
+    
     // setup
     @Before
     public void setUp() {
@@ -27,19 +28,25 @@ public class TestrunFilterDescriptorTest {
         value = "testValue";
         adminCases = true;
         activeTestcases = true;
-
+        
         this.testObject = new TestrunFilter(name, value, adminCases, activeTestcases);
         testObjectDescriptor = testObject.getDescriptor();
     }
-
+    
     // cleanup
     @After
     public void tearDown() {
         this.testObject = null;
     }
-
+    
     @Test
     public void getDisplayName() {
         assertEquals("Testrun Filter", testObjectDescriptor.getDisplayName());
+    }
+    
+    @Test
+    public void testConstructor() {
+        TestrunFilter.DescriptorImpl descriptor = new TestrunFilter.DescriptorImpl(TestrunFilter.class);
+        assertNotNull(descriptor);
     }
 }
